refactor(theme): update theme mode context import paths and create new themeModeContext.js file
